automod: quick extraction of flagstore package.
Did not rename, so some names are stuttery (flagstore.FlagStore).
This can be fixed in a future pass if desired; today's goal is just
more packages and getting the filesystem tree approximately right.
