+++ automod/account_meta.go
@@ -10,6 +10,7 @@ import (
 	appbsky "github.com/bluesky-social/indigo/api/bsky"
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/bluesky-social/indigo/automod/util"
 )
 
 type ProfileSummary struct {
@@ -95,8 +96,8 @@ func (e *Engine) GetAccountMeta(ctx context.Context, ident *identity.Identity) (
 			Description: pv.Description,
 			DisplayName: pv.DisplayName,
 		},
-		AccountLabels:        dedupeStrings(labels),
-		AccountNegatedLabels: dedupeStrings(negLabels),
+		AccountLabels:        util.DedupeStrings(labels),
+		AccountNegatedLabels: util.DedupeStrings(negLabels),
 		AccountFlags:         flags,
 	}
 	if pv.PostsCount != nil {
