+++ automod/flagstore/flagstore_redis_test.go
@@ -1,4 +1,4 @@
-package automod
+package flagstore
 
 import (
 	"context"
@@ -7,11 +7,15 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestFlagStoreBasics(t *testing.T) {
+func TestRedisFlagStoreBasics(t *testing.T) {
+	t.Skip("live test, need redis running locally")
 	assert := assert.New(t)
 	ctx := context.Background()
 
-	fs := NewMemFlagStore()
+	fs, err := NewRedisFlagStore("redis://localhost:6379/0")
+	if err != nil {
+		t.Fail()
+	}
 
 	l, err := fs.Get(ctx, "test1")
 	assert.NoError(err)
@@ -23,8 +27,9 @@ func TestFlagStoreBasics(t *testing.T) {
 	assert.NoError(err)
 	assert.Equal(3, len(l))
 
-	assert.NoError(fs.Remove(ctx, "test1", []string{"red", "blue"}))
+	assert.NoError(fs.Remove(ctx, "test1", []string{"red", "blue", "orange"}))
 	l, err = fs.Get(ctx, "test1")
 	assert.NoError(err)
 	assert.Equal([]string{"green"}, l)
+	assert.NoError(fs.Remove(ctx, "test1", []string{"green"}))
 }
